Use boost.test header-only
