+++ alpakaConfig.cmake
@@ -282,7 +282,6 @@ IF(${ALPAKA_DEBUG} GREATER 1)
     MESSAGE(STATUS "Boost_MINOR_VERSION: ${Boost_MINOR_VERSION}")
     MESSAGE(STATUS "Boost_SUBMINOR_VERSION: ${Boost_SUBMINOR_VERSION}")
     MESSAGE(STATUS "Boost_LIB_DIAGNOSTIC_DEFINITIONS: ${Boost_LIB_DIAGNOSTIC_DEFINITIONS}")
-    MESSAGE(STATUS "Boost_LIBRARIES: ${Boost_LIBRARIES}")
     MESSAGE(STATUS "Boost cached:")
     MESSAGE(STATUS "Boost_INCLUDE_DIR: ${Boost_INCLUDE_DIR}")
     MESSAGE(STATUS "Boost_LIBRARY_DIR: ${Boost_LIBRARY_DIR}")
