+++ test/CMakeLists.txt
@@ -26,17 +26,15 @@ FIND_PACKAGE(alpaka REQUIRED)
 
 ADD_SUBDIRECTORY("common/")
 
-FIND_PACKAGE(Boost QUIET COMPONENTS unit_test_framework)
-IF(NOT Boost_UNIT_TEST_FRAMEWORK_FOUND)
-    MESSAGE(FATAL_ERROR "Required alpaka dependency Boost.Test could not be found!")
-ENDIF()
-
 # NVCC does not incorporate the COMPILE_OPTIONS of a target but only the CMAKE_CXX_FLAGS
 IF(ALPAKA_ACC_GPU_CUDA_ENABLE AND NOT ALPAKA_CUDA_COMPILER MATCHES "clang")
     GET_TARGET_PROPERTY(_COMMON_COMPILE_OPTIONS common COMPILE_OPTIONS)
     STRING(REPLACE ";" " " _COMMON_COMPILE_OPTIONS_STRING "${_COMMON_COMPILE_OPTIONS}")
     SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} ${_COMMON_COMPILE_OPTIONS_STRING}")
 ENDIF()
+
+LIST(APPEND _ALPAKA_TEST_OPTIONS "--log_level=test_suite")
+LIST(APPEND _ALPAKA_TEST_OPTIONS "--color_output=true")
 
 ADD_SUBDIRECTORY("analysis/")
 ADD_SUBDIRECTORY("integ/")
