+++ test/common/CMakeLists.txt
@@ -43,6 +43,11 @@ IF(ALPAKA_ACC_GPU_CUDA_ENABLE OR (ALPAKA_ACC_GPU_HIP_ENABLE AND HIP_PLATFORM MAT
     LIST(APPEND _COMMON_COMPILE_DEFINITIONS_PUBLIC "CUDA_API_PER_THREAD_DEFAULT_STREAM")
 ENDIF()
 
+IF(MSVC)
+    # Disable autolink feature of boost. We explicitly link the required libs.
+    LIST(APPEND _COMMON_COMPILE_DEFINITIONS_PUBLIC "BOOST_ALL_NO_LIB")
+ENDIF()
+
 ADD_LIBRARY(
     ${_COMMON_TARGET_NAME}
     STATIC
