+++ test/unit/kernel/CMakeLists.txt
@@ -30,13 +30,8 @@ TARGET_INCLUDE_DIRECTORIES(
     PRIVATE ${Boost_INCLUDE_DIRS})
 TARGET_LINK_LIBRARIES(
     ${_TARGET_NAME}
-    PRIVATE common ${Boost_LIBRARIES})
-IF(NOT Boost_USE_STATIC_LIBS)
-    TARGET_COMPILE_DEFINITIONS(
-        ${_TARGET_NAME}
-        PRIVATE "BOOST_TEST_DYN_LINK")
-ENDIF()
+    PRIVATE common)
 
 SET_TARGET_PROPERTIES(${_TARGET_NAME} PROPERTIES FOLDER "test/unit")
 
-ADD_TEST(NAME ${_TARGET_NAME} COMMAND ${_TARGET_NAME} --log_level=test_suite --color_output=true)
+ADD_TEST(NAME ${_TARGET_NAME} COMMAND ${_TARGET_NAME} ${_ALPAKA_TEST_OPTIONS})
